SAM_AI_v3/requirements.txt
flask
flask-cors
anthropic
openai
google-generativeai
PyPDF2
python-docx
